Two bugs made Android drop content permanently and silently
The QA satellite matrix failed on every leg involving Android, in both
directions, across many runs. It was two independent bugs.

1. A CRASH, every time the circle switcher opened.

   java.util.IllegalFormatConversionException: d != kotlin.UInt
     at CircleScreenKt$CircleSwitcher$1$4$1$1.invoke(CircleScreen.kt:1705)

   uniffi maps Rust u32 to kotlin.UInt, which BOXES as kotlin.UInt and not as
   Integer, so Android's %d formatter rejects it outright. stringResource takes
   `vararg Any`, so nothing warns at compile time — it just kills the app at
   render time. This is the reported "Haven is not responding on nearly every
   launch", and every post that arrived while it was dead went unread.

   All 101 parameterized format calls were swept; every other numeric argument
   resolves to Int, so this was the only live site. FormatArgTypeTest now scans
   the sources against the unsigned fields the generated bindings declare, so
   the next one fails a test instead of a launch.

2. CONTENT THAT ARRIVED AND WAS NEVER READABLE AGAIN.

   Android's engine state, pulled off the device and opened in the harness:

     shared circle:  events=1, parked=8
     rosters:        b2d704e8 v4 (3 devices), fe263256 v1 (1 device)

   Eight posts had ARRIVED. They sat in `pending_epoch`, the durable buffer for
   an envelope that shows up before the key that opens it or the roster that
   resolves its sender. The mailbox drain deliberately marks such a key SEEN —
   the bytes are already held, and re-fetching them forever is the re-ingest
   storm this codebase has fought before — so recovery depends entirely on
   something replaying that buffer later. Two defects removed every trigger:

     * ingest_roster_wire (the relay-PULL arm, and the ONLY way a contact's
       roster ever reaches Android) never drained, while the mailbox arm always
       had. Same bytes, same verification, different consequences.

     * re-ingesting a roster we ALREADY HOLD returned false — indistinguishable
       from a forgery or a rollback. Callers read that bool as "is this roster
       known?", so a healthy device logged "devroster INGEST REJECTED … the
       engine refused them" on every poll and skipped its post-roster recovery
       forever.

   Now the pull arm drains like the mailbox arm, and already-current reports
   success. The rollback defense is unchanged and separately pinned: a genuinely
   OLDER roster is still refused.

   This was never Android-specific. ingest_roster_wire is shared, so any client
   whose contact roster arrives by pull could lose content the same way.

And the reason this took so long to see: nothing could tell "never arrived" from
"arrived and could not be opened". Those look identical from a feed — a short
list and no log line — and they have opposite fixes. diag_delivery_json reports
parked depth, epoch keys held and known rosters, and it is now in all four QA
dumps, so the next failure states which one it is instead of being inferred from
logcat and a hand-decoded state file.

core: 97 tests green. Android unit tests green. Desktop compiles.
Regression tests verified to FAIL without these fixes.

+++ android/app/src/main/java/com/blaineam/haven/ui/CircleScreen.kt
@@ -1702,7 +1702,10 @@ private fun CircleSwitcher(activeId: String, circlesVersion: Int, modifier: Modi
                 androidx.compose.material3.DropdownMenuItem(
                     // Through circleName, not c.name — the switcher is a display site, so a circle
                     // I've privately renamed must read the same here as it does in the header.
-                    text = { Text(stringResource(R.string.circle_name_member_count, HavenNet.circleName(c.id), c.memberCount), color = HavenTheme.textPrimary) },
+                    // memberCount is a uniffi u32 -> kotlin.UInt, which BOXES as kotlin.UInt and not as
+                    // Integer. Android's %d formatter rejects that with IllegalFormatConversionException,
+                    // so rendering this row crashed the app outright every time the switcher opened.
+                    text = { Text(stringResource(R.string.circle_name_member_count, HavenNet.circleName(c.id), c.memberCount.toInt()), color = HavenTheme.textPrimary) },
                     onClick = { HavenNet.setActiveCircle(c.id); menu = false },
                 )
             }
